processed-topics-consumer-container/requirements.txt
kafka-python
mysql-connector-python
psycopg2-binary
redis
pymongo
elasticsearch
